+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.40.0] - 2020-11-04
+## [0.33.0] - 2020-11-04
 ### Changes
 - BREAKING: Rename feature 'serde' as 'serde_feature' (for technical reasons)
 - BREAKING: Consolidate module-wide errors into a crate-wide error module